Add GDNative configuration for `testlib` use. See below for why.
While making a `testlib.gdnlib` for the shared library which has
`foreigner.so` etc as the entries is a bit unusual I think it is
the most "tidy" approach for handling dependencies & a situation
where potentially there's multiple shared libraries in a single
project but you still want to be able to drag & drop the config
for a single library into a new project.

The `testlib` shared library binaries need to be dependencies so
that they are copied & included on project export.