Add updated shared library binaries.
These were all built from `foreigner`:

 * 38455861aee44be621f45280e5cd7404f36058a6 `workaround-tlv-and-terminator-dlclose-issues`

And built against `godot-cpp`:

 * 9eceb16f0553884094d0f659461649be5d333866 ~`master` ~`3.2`

The Win64 build of `godot-cpp` had an edit to `include/core/Defs.hpp`
corresponding to: <https://github.com/godotengine/godot-cpp/pull/415#issuecomment-649596453>

The primary change is that they are built against a `3.2`ish `godot-cpp`
which fixed `Ref<>` issues and thanks to that and some other work
they should now not leak memory or have errors/warnings on exit.

The `.so` & `.dll` are stripped of debug symbols because they were
*massively* increased since 3.1.

What is this "CI" you speak of? :D